Refactored to have a separate command line tool and file processing
